+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Admin/AdminWebApiConfig.cs
@@ -20,39 +20,16 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Newtonsoft.Json.Linq;
-using System;
-using System.Diagnostics.CodeAnalysis;
-using System.Threading;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.ApplicationParts;
+using Stormancer.Server.Plugins.AdminApi;
 
 namespace Stormancer.Server.Plugins.GameSession
 {
-    public class GameServerInstance
+    class AdminWebApiConfig : IAdminWebApiConfig
     {
-        public string Id { get; set; }
-        public Action OnClosed { get; set; }
-    }
-
-    public class StartGameServerResult
-    {
-        public StartGameServerResult(bool success, GameServerInstance? instance, object? context )
+        public void ConfigureApplicationParts(ApplicationPartManager apm)
         {
-            Success = success;
-            Instance = instance;
-            Context = context;
+            apm.ApplicationParts.Add(new AssemblyPart(this.GetType().Assembly));
         }
-
-        [MemberNotNullWhen(true,"Instance")]
-        public bool Success { get; }
-        public GameServerInstance? Instance { get; }
-        public object? Context { get; set; }
-    }
-    public interface IGameServerProvider
-    {
-        string Type { get; }
-        Task<StartGameServerResult> TryStartServer(string id, JObject config, CancellationToken ct);
-
-        Task StopServer(string id, object? context);
     }
 }