+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/GameSessionPlugin.cs
@@ -25,9 +25,12 @@ using Stormancer.Core;
 using Stormancer.Diagnostics;
 using Stormancer.Plugins;
 using Stormancer.Server.Components;
+using Stormancer.Server.Plugins.AdminApi;
 using Stormancer.Server.Plugins.Analytics;
 using Stormancer.Server.Plugins.Configuration;
+using Stormancer.Server.Plugins.GameSession.Admin;
 using Stormancer.Server.Plugins.GameSession.ServerPool;
+using Stormancer.Server.Plugins.GameSession.ServerProviders;
 using Stormancer.Server.Plugins.ServiceLocator;
 using Stormancer.Server.Plugins.Users;
 using System;
@@ -51,6 +54,8 @@ namespace Stormancer.Server.Plugins.GameSession
                 builder.Register<GameSessionController>().InstancePerRequest();
                 builder.Register<ServerPoolController>().InstancePerRequest();
                 builder.Register<DedicatedServerAuthProvider>().As<IAuthenticationProvider>();
+                builder.Register<GameServerAgentAuthenticationProvider>().As<IAuthenticationProvider>();
+                builder.Register<GameServerAgentConfiguration>().As<IConfigurationChangedEventHandler>().AsSelf().SingleInstance();
                 builder.Register<DevDedicatedServerAuthProvider>().As<IAuthenticationProvider>();
                 builder.Register<GameSessions>().As<IGameSessions>();
                 builder.Register<ServerPools>().As<IServerPools>().AsSelf().As<IConfigurationChangedEventHandler>().InstancePerScene();
@@ -59,6 +64,10 @@ namespace Stormancer.Server.Plugins.GameSession
                 builder.Register<ProviderBasedServerPoolProvider>().As<IServerPoolProvider>().InstancePerScene();
                 builder.Register<DockerGameServerProvider>().As<IGameServerProvider>().SingleInstance();
                 builder.Register<GameSessionAnalyticsWorker>().SingleInstance();
+                builder.Register<AgentBasedGameServerProvider>().As<IGameServerProvider>().SingleInstance();
+                builder.Register<AdminWebApiConfig>().As<IAdminWebApiConfig>();
+                builder.Register<DockerAgentAdminController>();
+
             };
 
             ctx.HostStarting += (IHost host) =>